CC-64 Refacto APIs: Move city and bus gets to public, add get all trips to backoffice
